[Award] レビュー対応: AwardChip 共通化・awardName の trim 追加
- _AwardChip の重複定義を lib/presentation/widget/award_chip.dart に共通化
- awardDisplayText getter で awardName に trim() を適用し空白のみのケースを除外
