+++ lib/features/research_work/presentation/widgets/work_info_section.dart
@@ -4,6 +4,7 @@ import "package:kenryo_tankyu/core/constants/work/category_value.dart";
 import "package:kenryo_tankyu/core/constants/work/info_value.dart";
 import "package:kenryo_tankyu/core/constants/work/sub_category_value.dart";
 import 'package:kenryo_tankyu/features/research_work/domain/models/searched.dart';
+import 'package:kenryo_tankyu/presentation/widget/award_chip.dart';
 
 class WorkInfoSection extends StatelessWidget {
   final Searched searched;
@@ -86,7 +87,7 @@ class WorkDetailsTable extends StatelessWidget {
                 ),
               ),
               if (searched.awardDisplayText != null)
-                _AwardChip(text: searched.awardDisplayText!),
+                AwardChip(text: searched.awardDisplayText!),
             ],
           ),
         ),
@@ -113,39 +114,6 @@ class WorkDetailsTable extends StatelessWidget {
           ),
         ),
       ],
-    );
-  }
-}
-
-class _AwardChip extends StatelessWidget {
-  final String text;
-  const _AwardChip({required this.text});
-
-  @override
-  Widget build(BuildContext context) {
-    const color = Color(0xFFB8860B);
-    return Container(
-      padding: const EdgeInsets.symmetric(horizontal: 6, vertical: 2),
-      decoration: BoxDecoration(
-        color: color.withValues(alpha: 0.1),
-        borderRadius: BorderRadius.circular(4),
-        border: Border.all(color: color.withValues(alpha: 0.5)),
-      ),
-      child: Row(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          const Icon(Icons.emoji_events_outlined, size: 11, color: color),
-          const SizedBox(width: 3),
-          Text(
-            text,
-            style: const TextStyle(
-              fontSize: 11,
-              color: color,
-              fontWeight: FontWeight.bold,
-            ),
-          ),
-        ],
-      ),
     );
   }
 }
