+++ lib/presentation/widget/award_chip.dart
@@ -1,0 +1,34 @@
+import 'package:flutter/material.dart';
+
+class AwardChip extends StatelessWidget {
+  final String text;
+  const AwardChip({super.key, required this.text});
+
+  @override
+  Widget build(BuildContext context) {
+    const color = Color(0xFFB8860B);
+    return Container(
+      padding: const EdgeInsets.symmetric(horizontal: 6, vertical: 2),
+      decoration: BoxDecoration(
+        color: color.withValues(alpha: 0.1),
+        borderRadius: BorderRadius.circular(4),
+        border: Border.all(color: color.withValues(alpha: 0.5)),
+      ),
+      child: Row(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          const Icon(Icons.emoji_events_outlined, size: 11, color: color),
+          const SizedBox(width: 3),
+          Text(
+            text,
+            style: const TextStyle(
+              fontSize: 11,
+              color: color,
+              fontWeight: FontWeight.bold,
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
